+++ src/constant_time.rs
@@ -32,7 +32,7 @@ pub fn verify_slices_are_equal(a: &[u8], b: &[u8]) -> Result<(), error::Unspecif
 }
 
 prefixed_extern! {
-    fn OPENSSL_memcmp(a: *const u8, b: *const u8, len: c::size_t) -> c::int;
+    fn OPENSSL_memcmp(a: *const u8, b: *const u8, len: c::size_t) -> c::size_t;
 }
 
 #[cfg(test)]
